move to CalculatorLibrary API
